// src/api/model/mapList.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * API macarte
 * OpenAPI spec version: 1.0
 */
import type { Bbox } from './bbox';

/**
 * Carte
 */
export interface MapList {
  title?: string;
  /** Titre de la carte, adapté pour etre utilisé dans les url */
  title_url?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  theme?: string | null;
  /** @nullable */
  theme_id?: number | null;
  /** identifiant de l'organisation */
  organization_id?: string;
  /** Nom de l'organisation */
  organization_name?: string;
  type?: string;
  premium?: string;
  created_at?: string;
  updated_at?: string;
  nb_view?: number;
  /** @nullable */
  img_url?: string | null;
  share?: string;
  bbox?: Bbox;
  view_id?: string;
  /** lien vers la visualisation de la carte */
  view_url?: string;
  /** si auteur de la carte */
  edit_id?: string;
  author?: string;
}