// src/api/model/mapResearch.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * API macarte
 * OpenAPI spec version: 1.0
 */
import type { MapList } from './mapList';
import type { MapResearchActivesItem } from './mapResearchActivesItem';
import type { MapResearchPremiumsItem } from './mapResearchPremiumsItem';
import type { MapResearchSharesItem } from './mapResearchSharesItem';
import type { MapResearchTypesItem } from './mapResearchTypesItem';
import type { MapResearchUsersItem } from './mapResearchUsersItem';
import type { MapResearchValidesItem } from './mapResearchValidesItem';
import type { Theme } from './theme';

/**
 * Recherche des cartes
 */
export interface MapResearch {
  maps?: MapList[];
  themes?: Theme[];
  users?: MapResearchUsersItem[];
  types?: MapResearchTypesItem[];
  premiums?: MapResearchPremiumsItem[];
  actives?: MapResearchActivesItem[];
  valides?: MapResearchValidesItem[];
  shares?: MapResearchSharesItem[];
  /** Mot(s) contenu(s) dans le titre, la description ou le theme des cartes */
  query?: string;
  /** Nombre de cartes correspondant à la recherche */
  count?: number;
  /** Nombre de cartes recues dans la requete */
  limit?: number;
  /** Nombre de cartes à passer avant de ls inclure dans la requete */
  offset?: number;
}